app/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
requests
prometheus-fastapi-instrumentator